backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
supabase
httpx
geopy
numpy
scikit-learn
